--- a/spec/models/service_spec.rb
+++ b/spec/models/service_spec.rb
@@ -1,4 +1,6 @@
 describe Service do
+  include_examples "miq ownership"
+
   context "service events" do
     before(:each) do
       @service = FactoryGirl.create(:service)