Purpose or Intent
-----------------
Speeds up "VMs & Instances in My LDAP Group" index and ajax calls in the `/vm_or_template` (Services -> Workload) section of the application, and does so without needing to affect controller or view code.


How
---
This adds `virtual_attribute` on to the `ownership_mixin`, which is used by a few classes and makes it so that MiqExpressions will evaluate filter in SQL instead of pulling down all matching records to then iterate over and filter in Ruby (causes a Massive N+1 on large datasets).  This does not require a join for the `virtual_attribute` to function, but in doing so, we lose a bit of possible performance gains.  That said, this makes the code much more flexible by skipping the join, so this is preferred for now.

The code changes that make use of a JOIN can be [found here](https://github.com/ManageIQ/manageiq/pull/10290) that you can use as a way of comparing the code change footprints.


Metrics
-------
The data set for these metrics used a DB with approximately 20k `Vm` records.


**Before (Regular User)**

Started POST "/vm_or_template/tree_select/?id=:id"...
Completed 200 OK in 7242ms (Views: 2.0ms | ActiveRecord: 884.1ms)

|      ms |   bytes |   objects | queries | query (ms) |   rows |
|    ---: |    ---: |      ---: |    ---: |       ---: |   ---: |
| 6,043.1 |     N/A | 3,769,724 |      44 |    4,158.3 | 29,709 |


**After (Regular User)**

Started POST "/vm_or_template/tree_select/?id=:id"...
Completed 200 OK in 1092ms (Views: 4.1ms | ActiveRecord: 179.3ms)

|      ms |   bytes |   objects | queries | query (ms) |   rows |
|    ---: |    ---: |      ---: |    ---: |       ---: |   ---: |
|   903.0 |     N/A |   783,617 |      30 |      397.8 |    397 |


**Before (Admin)**

Started POST "/vm_or_template/tree_select/?id=:id"...
Completed 200 OK in 94316ms (Views: 1.1ms | ActiveRecord: 13351.1ms)

|       ms | bytes |    objects | queries | query (ms) |    rows |
|     ---: |  ---: |       ---: |    ---: |       ---: |    ---: |
| 93,770.4 |   N/A | 57,852,016 |     112 |   78,603.0 | 349,084 |



**After (Admin)**

Started POST "/vm_or_template/tree_select/?id=:id"...
Completed 200 OK in 636ms (Views: 2.6ms | ActiveRecord: 319.6ms)

|       ms | bytes |    objects | queries | query (ms) |    rows |
|     ---: |  ---: |       ---: |    ---: |       ---: |    ---: |
|    524.4 |   N/A |    125,591 |      20 |      634.3 |      14 |



Links
-----
* https://www.pivotaltracker.com/story/show/127472985
* https://github.com/ManageIQ/manageiq/pull/10290 (old attempt)