@kbrock 2016-08-23
@dmetzger57 think this is darga/yes, but no bz... your take?