@dmetzger57 2016-08-23
I'm going darga/no. Reasoning: while this is a fantastic performance improvement, reducing code churn in the errata releases to resolving app failures and reported issues (BZs) is taking presidence in my perspective for this one.