@dmetzger57 2016-09-06
@chessbyte would you make this a drag/yes and back port