@chessbyte 2016-09-08
Darga Backport details:
```sh
$ git cherry-pick -x -m 1 9f12f51 
[darga 9d4faf5] Merge pull request #10685 from NickLaMuro/owned_by_ldap_sql_virtual_attribute
 Author: Keenan Brock <keenan@thebrocks.net>
 Date: Mon Aug 22 23:54:58 2016 -0400
 5 files changed, 153 insertions(+), 1 deletion(-)
 create mode 100644 spec/support/examples_group/shared_examples_for_ownership_mixin.rb

$ git log
commit 9d4faf52c39cc5d9250ca1d630587b948d4c361e
Author: Keenan Brock <keenan@thebrocks.net>
Date:   Mon Aug 22 23:54:58 2016 -0400

    Merge pull request #10685 from NickLaMuro/owned_by_ldap_sql_virtual_attribute
    
    Take 2:  Speed up "VMs & Instances in My LDAP Group" filter in /vm_or_template/explorer
    (cherry picked from commit 9f12f510073a8347a7fdec3a339123e45e63f42c)
```