@JPrause 2016-09-12
https://bugzilla.redhat.com/show_bug.cgi?id=1375346